requirements.txt
bottle
bottle_session
bottle_sqlalchemy
bottle_react
bottle_beaker
psycopg2
bcrypt
pytvmaze
slugify
